Add Leaf and Protoleaf structs
Add a Protoleaf struct to represent the payload of an add-request call
to a sigsum log, with corresponding from_ascii() and to_ascii()
functions.

Add a Leaf struct as well, for completion. (Note that leaves are only
serialized in the get-leaves endpoint, where a single ascii document
reperesents multiple leaves. Figuring out a proper interface to do this
is TBD.)

Add missing to_ascii() functions for SignedTreeHead, InclusionProof and
SigsumSignature.